pyproject.toml
[project]
name = "wind-turbines"
version = "0.0.1"

dependencies = []

[project.optional-dependencies]
dev = [
    "pyspark==3.5.3",
    "delta-spark==3.3.0",
    "pytest==8.3.3",
    "pytest-cov==3.0.0",
    "flake8==7.1.1",
    "Flake8-pyproject==1.2.3",
    "flake8-pytest-style==2.1.0",
    "black==24.10.0",
    "mypy==1.13.0",
    "pre-commit==4.0.1",
    "isort==5.13.2",
    "chispa==0.10.0",
]

[project.scripts]
ingest_data = "wind_turbines.main:ingest_data"
clean_data = "wind_turbines.main:clean_data"
identify_anomalies = "wind_turbines.main:identify_anomalies"
calculate_summary_statistics = "wind_turbines.main:calculate_summary_statistics"

[tool.pytest.ini_options]
pythonpath = [
  "src/"
]

[tool.flake8]
max-line-length = 88

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
